Feat(Lexer): Cleanup LexerContext
- Move Unnecesary members from Lexer Context to Context
- Remove Support to Try and Forget now grammar lexically has to be LL(1)
- Matching folder structure to namespaces

+++ breadcpp/src/bread/token.hpp
@@ -15,7 +15,7 @@ namespace bread {
         StringLiteral,  // "string-literal"
 
 
-        OpAssign,       // =s
+        OpAssign,       // =
         OpEquals,       // ==
         OpNotEquals,    // !=
 
@@ -71,7 +71,7 @@ namespace bread {
     struct Token {
         TokenType type;
         Location begin;
-        int length;
+        int64_t length;
         StringType str = "";
     };
